+++ remove-extra-media.php
@@ -33,21 +33,22 @@ class Remove_Extra_Media {
 
 	public static $donate_button;
 	public static $menu_id;
+	public static $post_id;
 	public static $settings_link;
 
 
 	public function __construct() {
-		add_action( 'admin_init', array( $this, 'admin_init' ) );
-		add_action( 'admin_menu', array( $this, 'admin_menu' ) );
-		add_action( 'init', array( $this, 'init' ) );
+		add_action( 'admin_init', array( __CLASS__, 'admin_init' ) );
+		add_action( 'admin_menu', array( __CLASS__, 'admin_menu' ) );
+		add_action( 'init', array( __CLASS__, 'init' ) );
 		self::$base = plugin_basename( __FILE__ );
 	}
 
 
-	public function admin_init() {
-		$this->update();
-		add_filter( 'plugin_action_links', array( $this, 'plugin_action_links' ), 10, 2 );
-		add_filter( 'plugin_row_meta', array( $this, 'plugin_row_meta' ), 10, 2 );
+	public static function admin_init() {
+		self::update();
+		add_filter( 'plugin_action_links', array( __CLASS__, 'plugin_action_links' ), 10, 2 );
+		add_filter( 'plugin_row_meta', array( __CLASS__, 'plugin_row_meta' ), 10, 2 );
 
 		self::$donate_button = <<<EOD
 <form action="https://www.paypal.com/cgi-bin/webscr" method="post" target="_top">
@@ -62,11 +63,11 @@ EOD;
 	}
 
 
-	public function admin_menu() {
-		self::$menu_id = add_management_page( esc_html__( 'Remove Extra Media Processer', 'remove-extra-media' ), esc_html__( 'Remove Extra Media Processer', 'remove-extra-media' ), 'manage_options', self::ID, array( $this, 'user_interface' ) );
+	public static function admin_menu() {
+		self::$menu_id = add_management_page( esc_html__( 'Remove Extra Media Processer', 'remove-extra-media' ), esc_html__( 'Remove Extra Media Processer', 'remove-extra-media' ), 'manage_options', self::ID, array( __CLASS__, 'user_interface' ) );
 
-		add_action( 'admin_print_scripts-' . self::$menu_id, array( $this, 'scripts' ) );
-		add_action( 'admin_print_styles-' . self::$menu_id, array( $this, 'styles' ) );
+		add_action( 'admin_print_scripts-' . self::$menu_id, array( __CLASS__, 'scripts' ) );
+		add_action( 'admin_print_styles-' . self::$menu_id, array( __CLASS__, 'styles' ) );
 
 		add_screen_meta_link(
 			'rmem_settings_link',
@@ -78,14 +79,14 @@ EOD;
 	}
 
 
-	public function init() {
-		add_action( 'wp_ajax_ajax_process_post', array( $this, 'ajax_process_post' ) );
+	public static function init() {
+		add_action( 'wp_ajax_ajax_process_post', array( __CLASS__, 'ajax_process_post' ) );
 		load_plugin_textdomain( self::ID, false, 'remove-extra-media/languages' );
 		self::set_post_types();
 	}
 
 
-	public function plugin_action_links( $links, $file ) {
+	public static function plugin_action_links( $links, $file ) {
 		if ( $file == self::$base ) {
 			array_unshift( $links, self::$settings_link );
 
@@ -156,10 +157,10 @@ EOD;
 	 * @SuppressWarnings(PHPMD.ExitExpression)
 	 * @SuppressWarnings(PHPMD.Superglobals)
 	 */
-	public function user_interface() {
+	public static function user_interface() {
 		// Capability check
 		if ( ! current_user_can( 'manage_options' ) )
-			wp_die( $this->post_id, esc_html__( "Your user account doesn't have permission to access this.", 'remove-extra-media' ) );
+			wp_die( self::$post_id, esc_html__( "Your user account doesn't have permission to access this.", 'remove-extra-media' ) );
 
 ?>
 
@@ -174,8 +175,8 @@ EOD;
 			$posts_to_import = rmem_get_option( 'posts_to_import' );
 			$posts_to_import = explode( ',', $posts_to_import );
 			foreach ( $posts_to_import as $post_id ) {
-				$this->post_id = $post_id;
-				$this->ajax_process_post();
+				self::$post_id = $post_id;
+				self::ajax_process_post();
 			}
 
 			exit( __LINE__ . ':' . basename( __FILE__ ) . " DONE<br />\n" );
@@ -201,10 +202,10 @@ EOD;
 			}
 
 			$posts = implode( ',', $posts );
-			$this->show_status( $count, $posts );
+			self::show_status( $count, $posts );
 		} else {
 			// No button click? Display the form.
-			$this->show_greeting();
+			self::show_greeting();
 		}
 ?>
 	</div>
@@ -245,7 +246,7 @@ EOD;
 	}
 
 
-	public function show_greeting() {
+	public static function show_greeting() {
 ?>
 	<form method="post" action="">
 <?php wp_nonce_field( self::ID ); ?>
@@ -272,7 +273,7 @@ EOD;
 	 *
 	 * @SuppressWarnings(PHPMD.Superglobals)
 	 */
-	public function show_status( $count, $posts ) {
+	public static function show_status( $count, $posts ) {
 		echo '<p>' . esc_html__( 'Please be patient while this script run. This can take a while, up to a minute per post. Do not navigate away from this page until this script is done or the import will not be completed. You will be notified via this page when the import is completed.', 'remove-extra-media' ) . '</p>';
 
 		echo '<p>' . sprintf( esc_html__( 'Estimated time required to import is %1$s minutes.', 'remove-extra-media' ), ( $count / 12 ) ) . '</p>';
@@ -421,22 +422,22 @@ EOD;
 	 * @SuppressWarnings(PHPMD.ExitExpression)
 	 * @SuppressWarnings(PHPMD.Superglobals)
 	 */
-	public function ajax_process_post() {
+	public static function ajax_process_post() {
 		if ( ! rmem_get_option( 'debug_mode' ) ) {
 			error_reporting( 0 ); // Don't break the JSON result
 			header( 'Content-type: application/json' );
-			$this->post_id = intval( $_REQUEST['id'] );
+			self::$post_id = intval( $_REQUEST['id'] );
 		}
 
-		$post = get_post( $this->post_id );
+		$post = get_post( self::$post_id );
 		if ( ! $post || ! in_array( $post->post_type, self::$post_types )  )
-			die( json_encode( array( 'error' => sprintf( esc_html__( 'Failed Processing: %s is incorrect post type.', 'remove-extra-media' ), esc_html( $this->post_id ) ) ) ) );
+			die( json_encode( array( 'error' => sprintf( esc_html__( 'Failed Processing: %s is incorrect post type.', 'remove-extra-media' ), esc_html( self::$post_id ) ) ) ) );
 
-		$count_removed = $this->do_remove_extra_media( $this->post_id, $post );
+		$count_removed = self::do_remove_extra_media( self::$post_id, $post );
 		if ( empty( $count_removed ) )
 			$count_removed = __( 'No', 'remove-extra-media' );
 
-		die( json_encode( array( 'success' => sprintf( __( '&quot;<a href="%1$s" target="_blank">%2$s</a>&quot; Post ID %3$s was successfully processed in %4$s seconds. <strong>%5$s</strong> extra media removed.', 'remove-extra-media' ), get_permalink( $this->post_id ), esc_html( get_the_title( $this->post_id ) ), $this->post_id, timer_stop(), $count_removed ) ) ) );
+		die( json_encode( array( 'success' => sprintf( __( '&quot;<a href="%1$s" target="_blank">%2$s</a>&quot; Post ID %3$s was successfully processed in %4$s seconds. <strong>%5$s</strong> extra media removed.', 'remove-extra-media' ), get_permalink( self::$post_id ), esc_html( get_the_title( self::$post_id ) ), self::$post_id, timer_stop(), $count_removed ) ) ) );
 	}
 
 
@@ -445,7 +446,7 @@ EOD;
 	 *
 	 * @SuppressWarnings(PHPMD.UnusedFormalParameter)
 	 */
-	public function do_remove_extra_media( $post_id, $post ) {
+	public static function do_remove_extra_media( $post_id, $post ) {
 		global $wpdb;
 
 		$featured_id = get_post_thumbnail_id( $post_id );
@@ -485,7 +486,7 @@ EOD;
 	}
 
 
-	public function admin_notices_0_0_1() {
+	public static function admin_notices_0_0_1() {
 		$content  = '<div class="updated fade"><p>';
 		$content .= sprintf( __( 'If your Remove Extra Media display has gone to funky town, please <a href="%s">read the FAQ</a> about possible CSS fixes.', 'remove-extra-media' ), 'https://aihrus.zendesk.com/entries/23722573-Major-Changes-Since-2-10-0' );
 		$content .= '</p></div>';
@@ -494,7 +495,7 @@ EOD;
 	}
 
 
-	public function admin_notices_donate() {
+	public static function admin_notices_donate() {
 		$content  = '<div class="updated fade"><p>';
 		$content .= sprintf( __( 'Please donate $5 towards development and support of this Remove Extra Media plugin. %s', 'remove-extra-media' ), self::$donate_button );
 		$content .= '</p></div>';
@@ -503,11 +504,11 @@ EOD;
 	}
 
 
-	public function update() {
+	public static function update() {
 		$prior_version = rmem_get_option( 'admin_notices' );
 		if ( $prior_version ) {
 			if ( $prior_version < '0.0.1' )
-				add_action( 'admin_notices', array( $this, 'admin_notices_0_0_1' ) );
+				add_action( 'admin_notices', array( __CLASS__, 'admin_notices_0_0_1' ) );
 
 			rmem_set_option( 'admin_notices' );
 		}
@@ -515,7 +516,7 @@ EOD;
 		// display donate on major/minor version release
 		$donate_version = rmem_get_option( 'donate_version', false );
 		if ( ! $donate_version || ( $donate_version != self::VERSION && preg_match( '#\.0$#', self::VERSION ) ) ) {
-			add_action( 'admin_notices', array( $this, 'admin_notices_donate' ) );
+			add_action( 'admin_notices', array( __CLASS__, 'admin_notices_donate' ) );
 			rmem_set_option( 'donate_version', self::VERSION );
 		}
 	}